feat:同步android源码

+++ android/app/src/main/java/com/github/tvbox/osc/server/RemoteServer.java
@@ -1,5 +1,6 @@
 package com.github.tvbox.osc.server;
 
+import static com.github.tvbox.osc.util.RegexUtils.getPattern;
 import android.annotation.SuppressLint;
 import android.content.Context;
 import android.net.wifi.WifiManager;
@@ -7,16 +8,15 @@ import android.os.Environment;
 import android.util.Base64;
 
 import com.github.tvbox.osc.R;
-import com.github.tvbox.osc.api.ApiConfig;
 import com.github.tvbox.osc.base.App;
-import com.github.tvbox.osc.event.RefreshEvent;
+import com.github.tvbox.osc.api.ApiConfig;
 import com.github.tvbox.osc.event.ServerEvent;
 import com.github.tvbox.osc.util.FileUtils;
 import com.github.tvbox.osc.util.OkGoHelper;
-import com.github.tvbox.osc.util.Proxy;
+import com.google.gson.Gson;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
-
+import com.github.tvbox.osc.util.Proxy;
 import org.greenrobot.eventbus.EventBus;
 
 import java.io.BufferedOutputStream;
@@ -26,13 +26,10 @@ import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
-import java.io.UnsupportedEncodingException;
 import java.net.Inet4Address;
 import java.net.InetAddress;
 import java.net.NetworkInterface;
 import java.net.SocketException;
-import java.net.URLDecoder;
-import java.net.URLEncoder;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -59,9 +56,11 @@ public class RemoteServer extends NanoHTTPD {
     public static int serverPort = 9978;
     private boolean isStarted = false;
     private DataReceiver mDataReceiver;
+    private ArrayList < RequestProcess > getRequestList = new ArrayList < > ();
+    private ArrayList < RequestProcess > postRequestList = new ArrayList < > ();
+    private static final String PATTERN_ETH_STR = "^eth\\d+$";
+    private static final Pattern ETH_PATTERN = Pattern.compile(PATTERN_ETH_STR);
     public static String m3u8Content;
-    private ArrayList<RequestProcess> getRequestList = new ArrayList<>();
-    private ArrayList<RequestProcess> postRequestList = new ArrayList<>();
 
     public RemoteServer(int port, Context context) {
         super(port);
@@ -99,6 +98,7 @@ public class RemoteServer extends NanoHTTPD {
 
     private Response getProxy(Object[] rs){
         try {
+            if (rs[0] instanceof NanoHTTPD.Response) return (NanoHTTPD.Response) rs[0];
             int code = (int) rs[0];
             String mime = (String) rs[1];
             InputStream stream = rs[2] != null ? (InputStream) rs[2] : null;
@@ -122,7 +122,6 @@ public class RemoteServer extends NanoHTTPD {
             return NanoHTTPD.newFixedLengthResponse(Response.Status.INTERNAL_ERROR, NanoHTTPD.MIME_PLAINTEXT, "500");
         }
     }
-
     @Override
     public Response serve(IHTTPSession session) {
         EventBus.getDefault().post(new ServerEvent(ServerEvent.SERVER_CONNECTION));
@@ -132,13 +131,15 @@ public class RemoteServer extends NanoHTTPD {
                 fileName = fileName.substring(0, fileName.indexOf('?'));
             }
             if (session.getMethod() == Method.GET) {
-                for (RequestProcess process : getRequestList) {
+                for (RequestProcess process: getRequestList) {
                     if (process.isRequest(session, fileName)) {
                         return process.doResponse(session, fileName, session.getParms(), null);
                     }
                 }
                 if (fileName.equals("/proxy")) {
-                    Map<String, String> params = session.getParms();
+                    Map < String, String > params = session.getParms();
+                    params.putAll(session.getHeaders());
+                    params.put("request-headers", new Gson().toJson(session.getHeaders()));
                     if (params.containsKey("do")) {
                         Object[] rs = ApiConfig.get().proxyLocal(params);
                         return getProxy(rs);
@@ -174,45 +175,26 @@ public class RemoteServer extends NanoHTTPD {
                         rs = new byte[0];
                     }
                     return NanoHTTPD.newFixedLengthResponse(NanoHTTPD.Response.Status.OK, "application/dns-message", new ByteArrayInputStream(rs), rs.length);
-                } else if (fileName.startsWith("/push/")) {
-                    String url = fileName.substring(6);
-                    if (url.startsWith("b64:")) {
-                        try {
-                            url = new String(Base64.decode(url.substring(4), Base64.DEFAULT | Base64.URL_SAFE | Base64.NO_WRAP), "UTF-8");
-                        } catch (UnsupportedEncodingException e) {
-                            e.printStackTrace();
-                        }
-                    } else {
-                        url = URLDecoder.decode(url);
-                    }
-                    EventBus.getDefault().post(new RefreshEvent(RefreshEvent.TYPE_PUSH_URL, url));
-                    return NanoHTTPD.newFixedLengthResponse(NanoHTTPD.Response.Status.OK, NanoHTTPD.MIME_PLAINTEXT, "ok");    
-                }  else if (fileName.startsWith("/proxyM3u8")) {
-//                    com.github.tvbox.osc.util.LOG.i("echo-m3u8:"+m3u8Content);
+                } else if (fileName.equals("/m3u8")) {
                     return NanoHTTPD.newFixedLengthResponse(NanoHTTPD.Response.Status.OK, NanoHTTPD.MIME_PLAINTEXT, m3u8Content);
-                }
-                 else if (fileName.startsWith("/dash/")) {
+                } else if (fileName.startsWith("/dash/")) {
                     String dashData = App.getInstance().getDashData();
                     try {
                         String data = new String(Base64.decode(dashData, Base64.DEFAULT | Base64.NO_WRAP), "UTF-8");
-                        return NanoHTTPD.newFixedLengthResponse(
-                                Response.Status.OK,
-                                "application/dash+xml",
-                                data
-                        );
+                        return NanoHTTPD.newFixedLengthResponse(NanoHTTPD.Response.Status.OK, "application/dash+xml", data);
                     } catch (Throwable th) {
                         return NanoHTTPD.newFixedLengthResponse(NanoHTTPD.Response.Status.INTERNAL_ERROR, NanoHTTPD.MIME_PLAINTEXT, dashData);
                     }
                 }
             } else if (session.getMethod() == Method.POST) {
-                Map<String, String> files = new HashMap<String, String>();
+                Map < String, String > files = new HashMap < String, String > ();
                 try {
                     if (session.getHeaders().containsKey("content-type")) {
                         String hd = session.getHeaders().get("content-type");
                         if (hd != null) {
                             // cuke: 修正中文乱码问题
                             if (hd.toLowerCase().contains("multipart/form-data") && !hd.toLowerCase().contains("charset=")) {
-                                Matcher matcher = Pattern.compile("[ |\t]*(boundary[ |\t]*=[ |\t]*['|\"]?[^\"^'^;^,]*['|\"]?)", Pattern.CASE_INSENSITIVE).matcher(hd);
+                                Matcher matcher = getPattern("[ |\t]*(boundary[ |\t]*=[ |\t]*['|\"]?[^\"^'^;^,]*['|\"]?)", Pattern.CASE_INSENSITIVE).matcher(hd);
                                 String boundary = matcher.find() ? matcher.group(1) : null;
                                 if (boundary != null) {
                                     session.getHeaders().put("content-type", "multipart/form-data; charset=utf-8; " + boundary);
@@ -226,24 +208,23 @@ public class RemoteServer extends NanoHTTPD {
                 } catch (NanoHTTPD.ResponseException rex) {
                     return createPlainTextResponse(rex.getStatus(), rex.getMessage());
                 }
-                for (RequestProcess process : postRequestList) {
+                for (RequestProcess process: postRequestList) {
                     if (process.isRequest(session, fileName)) {
                         return process.doResponse(session, fileName, session.getParms(), files);
                     }
                 }
                 try {
-                    Map<String, String> params = session.getParms();
+                    Map < String, String > params = session.getParms();
                     if (fileName.equals("/upload")) {
                         String path = params.get("path");
-                        for (String k : files.keySet()) {
+                        for (String k: files.keySet()) {
                             if (k.startsWith("files-")) {
                                 String fn = params.get(k);
                                 String tmpFile = files.get(k);
                                 File tmp = new File(tmpFile);
                                 String root = Environment.getExternalStorageDirectory().getAbsolutePath();
                                 File file = new File(root + "/" + path + "/" + fn);
-                                if (file.exists())
-                                    file.delete();
+                                if (file.exists()) file.delete();
                                 if (tmp.exists()) {
                                     if (fn.toLowerCase().endsWith(".zip")) {
                                         unzip(tmp, root + "/" + path);
@@ -251,8 +232,7 @@ public class RemoteServer extends NanoHTTPD {
                                         FileUtils.copyFile(tmp, file);
                                     }
                                 }
-                                if (tmp.exists())
-                                    tmp.delete();
+                                if (tmp.exists()) tmp.delete();
                             }
                         }
                         return NanoHTTPD.newFixedLengthResponse(NanoHTTPD.Response.Status.OK, NanoHTTPD.MIME_PLAINTEXT, "OK");
@@ -264,8 +244,7 @@ public class RemoteServer extends NanoHTTPD {
                         if (!file.exists()) {
                             file.mkdirs();
                             File flag = new File(root + "/" + path + "/" + name + "/.tvbox_folder");
-                            if (!flag.exists())
-                                flag.createNewFile();
+                            if (!flag.exists()) flag.createNewFile();
                         }
                         return NanoHTTPD.newFixedLengthResponse(NanoHTTPD.Response.Status.OK, NanoHTTPD.MIME_PLAINTEXT, "OK");
                     } else if (fileName.equals("/delFolder")) {
@@ -329,12 +308,12 @@ public class RemoteServer extends NanoHTTPD {
         int ipAddress = wifiManager.getConnectionInfo().getIpAddress();
         if (ipAddress == 0) {
             try {
-                Enumeration<NetworkInterface> enumerationNi = NetworkInterface.getNetworkInterfaces();
+                Enumeration < NetworkInterface > enumerationNi = NetworkInterface.getNetworkInterfaces();
                 while (enumerationNi.hasMoreElements()) {
                     NetworkInterface networkInterface = enumerationNi.nextElement();
                     String interfaceName = networkInterface.getDisplayName();
-                    if (interfaceName.equals("eth0") || interfaceName.equals("wlan0")) {
-                        Enumeration<InetAddress> enumIpAddr = networkInterface.getInetAddresses();
+                    if (ETH_PATTERN.matcher(interfaceName).matches() || interfaceName.equals("wlan0")) {
+                        Enumeration < InetAddress > enumIpAddr = networkInterface.getInetAddresses();
                         while (enumIpAddr.hasMoreElements()) {
                             InetAddress inetAddress = enumIpAddr.nextElement();
                             if (!inetAddress.isLoopbackAddress() && inetAddress instanceof Inet4Address) {
@@ -383,7 +362,7 @@ public class RemoteServer extends NanoHTTPD {
             }
         });
         JsonArray result = new JsonArray();
-        for (File f : list) {
+        for (File f: list) {
             if (f.getName().startsWith(".")) {
                 if (f.getName().equals(".tvbox_folder")) {
                     info.addProperty("del", 1);
@@ -407,7 +386,7 @@ public class RemoteServer extends NanoHTTPD {
             destDir.mkdirs();
         }
         ZipFile zip = new ZipFile(zipFilePath);
-        Enumeration<ZipEntry> iter = (Enumeration<ZipEntry>) zip.entries();
+        Enumeration < ZipEntry > iter = (Enumeration < ZipEntry > ) zip.entries();
         while (iter.hasMoreElements()) {
             ZipEntry entry = iter.nextElement();
             InputStream is = zip.getInputStream(entry);
@@ -416,19 +395,16 @@ public class RemoteServer extends NanoHTTPD {
                 extractFile(is, filePath);
             } else {
                 File dir = new File(filePath);
-                if (!dir.exists())
-                    dir.mkdirs();
+                if (!dir.exists()) dir.mkdirs();
                 File flag = new File(dir + "/.tvbox_folder");
-                if (!flag.exists())
-                    flag.createNewFile();
+                if (!flag.exists()) flag.createNewFile();
             }
         }
     }
 
     void extractFile(InputStream inputStream, String destFilePath) throws Throwable {
         File dst = new File(destFilePath);
-        if (dst.exists())
-            dst.delete();
+        if (dst.exists()) dst.delete();
         BufferedOutputStream bos = new BufferedOutputStream(new FileOutputStream(destFilePath));
         byte[] bytesIn = new byte[2048];
         int len = inputStream.read(bytesIn);
